Extract fragment
